add elements to dashboard
