+++ src/components/Header/Header.styled.jsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const MainHeader = styled.div`
+  height: 68px;
+`;
