+++ src/components/Sidebar/Sidebar.styled.js
@@ -1,28 +1,25 @@
 import styled from "styled-components";
 
 export const SidebarContainer = styled.div`
-	width: 260px;
-	height: 100%;
+  width: 260px;
 
-	position: absolute;
-	top: 0;
-	left: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
 
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
+  padding: 24px;
 
-	padding: 24px;
-
-	background-color: #121212;
-
-	@media (max-width: 1439px) {
-		display: none;
-	}
+  background-color: #121212;
 
 	@media (max-width: 767px) {
 		width: 225px;
 
 		padding: 14px 14px 24px 14px;
 	}
+
+
+	
+  @media (max-width: 1439px) {
+    display: none;
+  }
 `;
