+++ src/pages/HomePage.jsx
@@ -1,15 +1,20 @@
 import { Header } from "../components/Header/Header";
 import { Sidebar } from "../components/Sidebar/Sidebar";
+import { Conteiner } from "../shared/Container/Conteiner";
+import { HeaderAndScreensWrapper, HomePageWrapper } from "./HomePage.styled";
 import ScreensPage from "./ScreensPage";
-
 
 const HomePage = () => {
   return (
-    <div>
-      <Header />
-      <Sidebar />
-      <ScreensPage/>
-    </div>
+    <Conteiner>
+      <HomePageWrapper>
+        <Sidebar />
+        <HeaderAndScreensWrapper>
+          <Header />
+          <ScreensPage />
+        </HeaderAndScreensWrapper>
+      </HomePageWrapper>
+    </Conteiner>
   );
 };
