+++ src/pages/HomePage.styled.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const HomePageWrapper = styled.div`
+  display: flex;
+  min-height: 100vh;
+  /* height: 100vh; */
+`;
+
+export const HeaderAndScreensWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+`;
