+++ src/shared/Container/Conteiner.js
@@ -3,22 +3,24 @@ import styled from "styled-components";
 export const Conteiner = styled.div`
   margin-left: auto;
   margin-right: auto;
-  padding-left: 20px;
-  padding-right: 20px;
+  /* padding-left: 20px;
+  padding-right: 20px; */
 
 
-@media screen and (max-width: 767px) {
+@media screen and (max-width: 767.9px) {
     max-width: 375px;
 }
 
 @media screen and (min-width: 768px) {
     width: 768px;
-    padding-left: 32px;
-    padding-right: 32px;
+    /* padding-left: 32px;
+    padding-right: 32px; */
 }
 
 @media screen and (min-width: 1440px) {
     width: 1440px;
-    padding-left: 24px;
-    padding-right: 24px;
+
+		/* overflow-x: hidden; */
+    /* padding-left: 24px;
+    padding-right: 24px; */
   }`
